feat: sdist support
still need to write some more tests
