+++ Cargo.toml
@@ -49,7 +49,7 @@ rattler_solve = { version = "0.16.2", default-features = false, features = ["res
 rattler_virtual_packages = { version = "0.16.2", default-features = false }
 regex = "1.10.2"
 reqwest = { version = "0.11.23", default-features = false }
-rip = { package = "rattler_installs_packages", version = "0.1.0", default-features = false }
+rip = { package = "rattler_installs_packages", version = "0.2.1", default-features = false }
 serde = "1.0.195"
 serde-untagged = "0.1.5"
 serde_json = "1.0.111"
