+++ src/environment.rs
@@ -5,12 +5,14 @@ use crate::{
 use miette::{Context, IntoDiagnostic, LabeledSpan};
 
 use crate::lock_file::lock_file_satisfies_project;
+use crate::project::manifest::SystemRequirements;
 use crate::project::virtual_packages::verify_current_platform_has_required_virtual_packages;
 use rattler::install::{PythonInfo, Transaction};
 use rattler_conda_types::{Platform, PrefixRecord, RepoDataRecord};
 use rattler_lock::CondaLock;
 use rattler_repodata_gateway::sparse::SparseRepoData;
 use rip::index::PackageDb;
+use rip::python_env::PythonLocation;
 use std::{io::ErrorKind, path::Path};
 
 /// Verify the location of the prefix folder is not changed so the applied prefix path is still valid.
@@ -177,8 +179,18 @@ pub async fn get_up_to_date_prefix(
     };
 
     if project.has_pypi_dependencies() {
+        let python_location = match &python_status {
+            PythonStatus::Changed { new, .. }
+            | PythonStatus::Unchanged(new)
+            | PythonStatus::Added { new } => {
+                PythonLocation::Custom(prefix.root().join(new.path.clone()))
+            }
+            PythonStatus::DoesNotExist | PythonStatus::Removed { .. } => PythonLocation::System,
+        };
+
         if update_lock_file {
-            lock_file = lock_file::update_lock_file_for_pypi(project, lock_file).await?;
+            lock_file =
+                lock_file::update_lock_file_for_pypi(project, lock_file, &python_location).await?;
         }
 
         if !no_install {
@@ -189,6 +201,8 @@ pub async fn get_up_to_date_prefix(
                 project.pypi_package_db()?,
                 &lock_file,
                 &python_status,
+                &python_location,
+                &project.system_requirements(),
             )
             .await?;
         }
@@ -203,6 +217,8 @@ pub async fn update_prefix_pypi(
     package_db: &PackageDb,
     lock_file: &CondaLock,
     status: &PythonStatus,
+    python_location: &PythonLocation,
+    system_requirements: &SystemRequirements,
 ) -> miette::Result<()> {
     // Remove python packages from a previous python distribution if the python version changed.
     install_pypi::remove_old_python_distributions(prefix, platform, status)?;
@@ -210,7 +226,15 @@ pub async fn update_prefix_pypi(
     // Install and/or remove python packages
     progress::await_in_progress(
         "updating python packages",
-        install_pypi::update_python_distributions(package_db, prefix, lock_file, platform, status),
+        install_pypi::update_python_distributions(
+            package_db,
+            prefix,
+            lock_file,
+            platform,
+            status,
+            python_location,
+            system_requirements,
+        ),
     )
     .await
 }
