+++ src/install_pypi.rs
@@ -8,16 +8,23 @@ use indicatif::ProgressBar;
 use itertools::Itertools;
 use miette::{IntoDiagnostic, WrapErr};
 
+use crate::consts::PROJECT_MANIFEST;
+use crate::project::manifest::SystemRequirements;
+use crate::pypi_marker_env::determine_marker_environment;
+use crate::pypi_tags::{is_python_record, project_platform_tags};
 use rattler_conda_types::Platform;
 use rattler_lock::{CondaLock, LockedDependency};
 use rip::artifacts::wheel::{InstallPaths, UnpackWheelOptions};
 use rip::artifacts::Wheel;
 use rip::index::PackageDb;
-use rip::python_env::{find_distributions_in_venv, uninstall_distribution, Distribution, WheelTag};
+use rip::python_env::{
+    find_distributions_in_venv, uninstall_distribution, Distribution, PythonLocation, WheelTag,
+};
+use rip::resolve::ResolveOptions;
 use rip::types::{
     Artifact, ArtifactHashes, ArtifactInfo, ArtifactName, Extra, NormalizedPackageName,
-    WheelFilename,
 };
+use rip::wheel_builder::WheelBuilder;
 use std::collections::HashSet;
 use std::path::Path;
 use std::str::FromStr;
@@ -34,6 +41,8 @@ pub async fn update_python_distributions(
     lock_file: &CondaLock,
     platform: Platform,
     status: &PythonStatus,
+    python_location: &PythonLocation,
+    system_requirements: &SystemRequirements,
 ) -> miette::Result<()> {
     let python_info = match status {
         PythonStatus::Changed { new, .. }
@@ -76,9 +85,45 @@ pub async fn update_python_distributions(
             python_packages,
         );
 
+    let conda_packages = lock_file
+        .get_conda_packages_by_platform(platform)
+        .into_diagnostic()?;
+
+    // Determine the python interpreter that is installed as part of the conda packages.
+    let python_record = conda_packages
+        .iter()
+        .find(|r| is_python_record(r))
+        .ok_or_else(|| miette::miette!("could not resolve pypi dependencies because no python interpreter is added to the dependencies of the project.\nMake sure to add a python interpreter to the [dependencies] section of the {PROJECT_MANIFEST}, or run:\n\n\tpixi add python"))?;
+
+    // Determine the environment markers
+    let marker_environment = determine_marker_environment(platform, python_record.as_ref())?;
+
+    // Determine the compatible tags
+    let compatible_tags =
+        project_platform_tags(platform, system_requirements, python_record.as_ref());
+
+    let wheel_builder = if let PythonLocation::Custom(_) = python_location {
+        Some(WheelBuilder::new(
+            package_db,
+            &marker_environment,
+            Some(&compatible_tags),
+            &ResolveOptions {
+                sdist_resolution: rip::resolve::SDistResolution::PreferWheels,
+                python_location: python_location.clone(),
+            },
+            package_db.cache_dir(),
+        ))
+    } else {
+        None
+    };
+    let wheel_builder = wheel_builder.as_ref();
+
     // Start downloading the python packages that we want in the background.
-    let (package_stream, package_stream_pb) =
-        stream_python_artifacts(package_db, python_distributions_to_install.clone());
+    let (package_stream, package_stream_pb) = stream_python_artifacts(
+        package_db,
+        python_distributions_to_install.clone(),
+        wheel_builder,
+    );
 
     // Remove python packages that need to be removed
     if !python_distributions_to_remove.is_empty() {
@@ -194,6 +239,7 @@ async fn install_python_distributions(
 fn stream_python_artifacts<'a>(
     package_db: &'a PackageDb,
     packages_to_download: Vec<&'a LockedDependency>,
+    wheel_builder: Option<&'a WheelBuilder<'a, 'a>>,
 ) -> (
     impl Stream<Item = miette::Result<(Option<String>, HashSet<Extra>, Wheel)>> + 'a,
     Option<ProgressBar>,
@@ -217,6 +263,7 @@ fn stream_python_artifacts<'a>(
 
     let stream_pb = pb.clone();
     let total_packages = packages_to_download.len();
+
     let download_stream = stream::iter(packages_to_download)
         .map(move |package| {
             let pb = stream_pb.clone();
@@ -237,8 +284,9 @@ fn stream_python_artifacts<'a>(
                     .with_context(|| {
                         format!("'{}' is not a valid python package name", &package.name)
                     })?;
-                let wheel_name = WheelFilename::from_filename(filename, &name)
-                    .expect("failed to convert filename to wheel filename");
+
+                let artifact_name = ArtifactName::from_filename(filename, &name)
+                    .expect("failed to convert filename to artifact name");
 
                 // Log out intent to install this python package.
                 tracing::info!("downloading python package {filename}");
@@ -246,7 +294,7 @@ fn stream_python_artifacts<'a>(
 
                 // Reconstruct the ArtifactInfo from the data in the lockfile.
                 let artifact_info = ArtifactInfo {
-                    filename: ArtifactName::Wheel(wheel_name),
+                    filename: artifact_name,
                     url: pip_package.url.clone(),
                     hashes: pip_package.hash.as_ref().map(|hash| ArtifactHashes {
                         sha256: hash.sha256().cloned(),
@@ -263,7 +311,7 @@ fn stream_python_artifacts<'a>(
 
                 // TODO: Maybe we should have a cache of wheels separate from the package_db. Since a
                 //   wheel can just be identified by its hash or url.
-                let wheel: Wheel = package_db.get_wheel(&artifact_info, None).await?;
+                let wheel: Wheel = package_db.get_wheel(&artifact_info, wheel_builder).await?;
 
                 // Update the progress bar
                 pb_task.finish().await;
